requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
biopython>=1.79
tqdm>=4.62.0